@@ -6,6 +6,11 @@
   "watch": {
     "test": "{src,test}/*.js"
   },
+  "jest": {
+    "setupFiles": [
+      "raf/polyfill"
+    ]
+  },
   "scripts": {
     "webpack": "webpack",
     "test": "jest --verbose && lab -v -l",
@@ -52,15 +57,17 @@
     "jest": "^21.0.0",
     "jquery": "^3.2.0",
     "npm-watch": "^0.2.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-scripts": "^1.0.0",
     "vision": "^4.1.1",
     "webpack": "^3.0.0"
   },
   "devDependencies": {
     "code": "^5.0.0",
-    "enzyme": "^2.7.1",
+    "enzyme": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.0.2",
     "eslint": "^4.6.0",
     "eslint-config-eta": "^1.0.0",
     "eslint-plugin-babel": "^4.1.1",
@@ -68,6 +75,7 @@
     "jest-cli": "^21.0.0",
     "lab": "^14.0.0",
     "prettier": "^1.0.1",
+    "raf": "^3.4.0",
     "react-addons-test-utils": "^15.4.2",
     "react-test-renderer": "^15.4.2"
   }
